Move helper function for visits to visits/helpers
